% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_kphon_audio.R
\name{copy_kphon_audio}
\alias{copy_kphon_audio}
\title{Copy exported kphon audio to an external directory}
\usage{
copy_kphon_audio(kphon_df, kphon_path, dest_path, overwrite = TRUE)
}
\arguments{
\item{kphon_df}{subsetted kphon data frame with \code{kphon_export} column}

\item{kphon_path}{local path to kphon directory in which 'audio' directory exists}

\item{dest_path}{path to external directory}

\item{overwrite}{overwrite existing files in \code{dest_path}. Default is \code{TRUE}.}
}
\description{
Copy exported kphon audio to an external directory
}
