% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_kphon_wav.R
\name{hash_kphon_wav}
\alias{hash_kphon_wav}
\title{Create unique hash for kphon wav file}
\usage{
hash_kphon_wav(orthography, speaker, annotation_source, xmin)
}
\arguments{
\item{orthography}{orthographic represention of Kaytetye form}

\item{speaker}{initials of speaker in wav file}

\item{annotation_source}{TextGrid file from which annotation was extracted}

\item{xmin}{the start time of the annotation in TextGrid file}
}
\description{
Create unique hash for kphon wav file
}
\examples{

hash_kphon_wav("aleke", "AR", "southwindpassage_AR_MP.TextGrid", 4.56)

}
