% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_with_kdict.R
\name{join_with_kdict}
\alias{join_with_kdict}
\title{Left-join a data frame with Kaytetye dictionary data, based on column 'orthography' (default)}
\usage{
join_with_kdict(df, by = "orthography")
}
\arguments{
\item{df}{a data frame}

\item{by}{common column between \code{df} and KDB data frame}
}
\description{
Left-join a data frame with Kaytetye dictionary data, based on column 'orthography' (default)
}
