% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdb_to_kphon.R
\name{kdb_to_kphon}
\alias{kdb_to_kphon}
\title{Convert the Kaytetye dictionary database text file into a data frame for use within kphon}
\usage{
kdb_to_kphon(kdb_txt)
}
\arguments{
\item{kdb_txt}{path to KDB.txt file}
}
\description{
Convert the Kaytetye dictionary database text file into a data frame for use within kphon
}
