% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_kphon_master.R
\name{make_kphon_master}
\alias{make_kphon_master}
\title{Make kphon master file}
\usage{
make_kphon_master(kphon_path = ".", output_csv = FALSE)
}
\arguments{
\item{kphon_path}{local path to kphon directory in which 'annotations' directory exists}

\item{output_csv}{(over)write to _kphon-master.csv file? Default is false.}
}
\description{
Make kphon master file
}
