% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kphon_master.R
\name{read_kphon_master}
\alias{read_kphon_master}
\title{Read kphon master file}
\usage{
read_kphon_master(kphon_path = ".")
}
\arguments{
\item{kphon_path}{local path to kphon directory in which 'annotations' directory exists. Default is "."}
}
\description{
Read kphon master file
}
