% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_kphon_audio.R
\name{sync_kphon_audio}
\alias{sync_kphon_audio}
\title{Sync master audio directory with audio from all _kphon-export folders}
\usage{
sync_kphon_audio(kphon_path, show_summary = TRUE)
}
\arguments{
\item{kphon_path}{path to kphon directory where _kphon-master.csv file and annotations directory exists}

\item{show_summary}{return summary (default is \code{TRUE}). Must explicitly supply \code{FALSE} to begin moving new files and deleting stale files.}
}
\description{
Sync master audio directory with audio from all _kphon-export folders
}
