% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textgrid_to_kphon.R
\name{textgrid_to_kphon}
\alias{textgrid_to_kphon}
\title{Convert data from TextGrid file(s) to the kphon data frame specification}
\usage{
textgrid_to_kphon(input_files, io_mapping)
}
\arguments{
\item{input_files}{a vector of TextGrid files}

\item{io_mapping}{a list specifying input/output mapping configuration (see main README for specification)}
}
\description{
Convert data from TextGrid file(s) to the kphon data frame specification
}
