% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_kphon_exports.R
\name{write_kphon_exports}
\alias{write_kphon_exports}
\title{Test that latin/scientific name(s) lines are well-formed}
\usage{
write_kphon_exports(kphon_df, path, overwrite = FALSE,
  ffmpeg_path = "ffmpeg")
}
\arguments{
\item{kphon_df}{a kphon data frame, e.g. created using \code{textgrid_to_kphon}}

\item{path}{directory in which output files will be written}

\item{overwrite}{should existing wav files be overwritten (default is FALSE)}

\item{ffmpeg_path}{path to executable ffmpeg binary}
}
\description{
Test that latin/scientific name(s) lines are well-formed
}
